Split out special rules into new function for smart playlists
